Add support for Elasticsearch sniffer
Closes gh-24174
