+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/elasticsearch/ElasticsearchRestClientAutoConfigurationTests.java
@@ -34,12 +34,14 @@ import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestClient;
 import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
+import org.elasticsearch.client.sniff.Sniffer;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.elasticsearch.ElasticsearchContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.boot.testsupport.testcontainers.DockerImageNames;
 import org.springframework.context.annotation.Bean;
@@ -47,6 +49,7 @@ import org.springframework.context.annotation.Configuration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verifyNoInteractions;
 
 /**
  * Tests for {@link ElasticsearchRestClientAutoConfiguration}.
@@ -217,6 +220,52 @@ class ElasticsearchRestClientAutoConfigurationTests {
 								assertThat(defaultCredentials.getUserPrincipal().getName()).isEqualTo("admin");
 								assertThat(defaultCredentials.getPassword()).isEqualTo("admin");
 							});
+				});
+	}
+
+	@Test
+	void configureWithoutSnifferLibraryShouldNotCreateSniffer() {
+		this.contextRunner.withClassLoader(new FilteredClassLoader("org.elasticsearch.client.sniff"))
+				.run((context) -> assertThat(context).hasSingleBean(RestHighLevelClient.class)
+						.doesNotHaveBean(Sniffer.class));
+	}
+
+	@Test
+	void configureShouldCreateSnifferUsingRestHighLevelClient() {
+		this.contextRunner.run((context) -> {
+			assertThat(context).hasSingleBean(Sniffer.class);
+			assertThat(context.getBean(Sniffer.class)).hasFieldOrPropertyWithValue("restClient",
+					context.getBean(RestHighLevelClient.class).getLowLevelClient());
+			// Validate shutdown order as the sniffer must be shutdown before the client
+			assertThat(context.getBeanFactory().getDependentBeans("elasticsearchRestHighLevelClient"))
+					.contains("elasticsearchSniffer");
+		});
+	}
+
+	@Test
+	void configureWithCustomSnifferSettings() {
+		this.contextRunner.withPropertyValues("spring.elasticsearch.rest.sniffer.interval=180s",
+				"spring.elasticsearch.rest.sniffer.delay-after-failure=30s").run((context) -> {
+					assertThat(context).hasSingleBean(Sniffer.class);
+					Sniffer sniffer = context.getBean(Sniffer.class);
+					assertThat(sniffer).hasFieldOrPropertyWithValue("sniffIntervalMillis",
+							Duration.ofMinutes(3).toMillis());
+					assertThat(sniffer).hasFieldOrPropertyWithValue("sniffAfterFailureDelayMillis",
+							Duration.ofSeconds(30).toMillis());
+				});
+	}
+
+	@Test
+	void configureWhenCustomSnifferShouldBackOff() {
+		Sniffer customSniffer = mock(Sniffer.class);
+		this.contextRunner.withBean(Sniffer.class, () -> customSniffer)
+				.withPropertyValues("spring.elasticsearch.rest.sniffer.interval=180s",
+						"spring.elasticsearch.rest.sniffer.delay-after-failure=30s")
+				.run((context) -> {
+					assertThat(context).hasSingleBean(Sniffer.class);
+					Sniffer sniffer = context.getBean(Sniffer.class);
+					assertThat(sniffer).isSameAs(customSniffer);
+					verifyNoInteractions(customSniffer);
 				});
 	}
